The palette scrolls, continuously, and its colours travel with it
scrollSteps has been sitting in the state struct since the layout was written, read by absolutely
nothing. The palette never scrolled: seven fixed degrees, with octave as a separate discrete jump.
That is two ideas the user has to hold at once, and the relationship between them was never
explained anywhere because there was nothing sensible to say.

Now it is one motion. The horizontal wheel scrolls the palette, and it keeps going - past the top
of the scale and into the next octave, so the columns slide sideways through the whole range the
way the iso already slides vertically. Cody: "the palette to live scroll along ... move
horizontally across the octave so the columns move to the left and right like the chord mode
keyboard". The chord-tone maths already wrapped octaves correctly; only the column-to-degree
mapping was pinned, so this is mostly deleting an assumption.

And the colours now follow the DEGREE rather than the column. That is the half that makes it
legible: colouring by column would slide the chords while the colours stayed nailed to the grid,
so the palette would look motionless while its contents changed underneath - the most confusing
possible combination. Following the degree means you SEE it move.

Every column-to-degree conversion goes through degreeAtColumn() so the play path, the LEARN path
and the render path cannot drift apart.

Device build clean, 119 tests pass. Scroll bounds come from where chords would run off the ends of
MIDI, not an arbitrary octave count.

+++ src/deluge/gui/ui/keyboard/layout/harmonic.h
@@ -64,7 +64,7 @@ private:
 	static constexpr int32_t kBlockW = 7;       // each grid block is 7 columns wide
 
 	uint8_t getScaleIntervals(uint8_t* ivOut);
-	uint8_t buildChordAtDegree(uint8_t deg, int32_t y, const uint8_t* iv, uint8_t sc, uint8_t keyRoot,
+	uint8_t buildChordAtDegree(int32_t deg, int32_t y, const uint8_t* iv, uint8_t sc, uint8_t keyRoot,
 	                           int16_t* notesOut, uint8_t maxNotes, uint8_t* rootPcOut, char* romanOut, char* absOut);
 	int32_t isoNoteAt(int32_t localX, int32_t y);        // in-key mapping (localX = iso column 0..6)
 	int32_t isoNoteChromatic(int32_t localX, int32_t y); // chromatic isomorphic mapping (localX 0..6)
